Reverting to classic tmux colours
I think the green footer is iconic to tmux so let's gp with it instead
of trying to be overly stylish.

I also noticed that \ (i.e. | on the keyboard) seem to have stopped
splitting the pane horizontally.  No idea when or why this is the case,
but wraping in single quotes seems to have fixed it and I've also
escaped it for good meaure.  Wrapped '-' too for consitency.